Update SleepWithFriendsOneMinute.lua
local serverDayLength = (SandboxVars.DayLength - 2) added as proof of concept pulling server daytime setting. the -2 is to simplify math, as 1 hour is option 3, 2 hours option 4, and so on. Basically, forget 30 and 15 min servers lmao

added local AutoWake = SandboxVars.SleepWithFriends.AutoWake to set as new variable. Likely redundant and can be removed if found so.

added AutoWake == 1 to enable toggle